Move Domain Objects to Aggregate namespace
