+++ src/EightCare.Domain/KeeperAggregate/Animal.cs
@@ -1,4 +1,4 @@
-namespace EightCare.Domain
+namespace EightCare.Domain.KeeperAggregate
 {
     public sealed class Animal
     {
